refactor: point 조회 함수 order API에 포함
