+++ src/components/Cart/EstimatedPaymentBox.tsx
@@ -1,6 +1,6 @@
 import type { ScheduledOrderType } from '../../types/product';
 
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
